paths and turn system
THIS TOOK FOREVER OMG
but it's finally done. so i added paths which you can press q or e to choose a path but it's not visibly shown yet. and there is now a turn system but the ai is still not there so i will add something close to ai lol

-Max signing out...