improved the camera and made some pixel art
made an intro to the board where the camera moves towards the board then starts spinning like crazy. and also the board now has a name "Domain of Marcos"! i like it! and it has some pixel art to go along with it. also added a main menu and pixel art for the title of the game "Knowbody Nose"

~Max signing out...